Unify GitHub modules
